feat: initialize the google login callback
